refactor: :art: Applied code review requests.
