--- src/interfaces/IApiUserResponse.tsx
@@ -1,9 +1,0 @@
-import { User } from "./IUser";
-
-export interface ApiUserResponse {
-    isSucceed: boolean,
-    result: {
-        user: User
-    }
-    errorMessage?: string;
-}
