+++ src/services/userService.tsx
@@ -1,11 +1,11 @@
 import { OperationResult } from "../interfaces/IOperationResult";
 import { profileUpdateForm } from "../interfaces/IProfileUpdateForm";
 import { Fetcher } from "./fetcher";
-import { ApiUserResponse } from "../interfaces/IApiUserResponse";
+import { User } from "../interfaces/IUser";
 
 export class UserService extends Fetcher {
-    async updateProfile(values:profileUpdateForm, id:string):Promise<OperationResult<ApiUserResponse>> {
-        const response = await super.patch<profileUpdateForm, ApiUserResponse>(`/users/${id}`, values);
+    async updateProfile(values:profileUpdateForm, id:string):Promise<OperationResult<User>> {
+        const response = await super.patch<profileUpdateForm, User>(`/users/${id}`, values);
         return response.data;
     }
 }
